agent-test/requirements.txt
anms-ace ~= 1.0
cbor2 ~= 5.6
pytest
